Florian feature

+++ backend/PrimeTree/src/main/java/BackendServer/User/Service/UserManager.java
@@ -20,7 +20,7 @@ public interface UserManager extends UserDetailsService{
 	 * @return the favourite list of the user
 	 * @throws UserNotFoundException if no user with id userId exists
 	 */
-	public int[] getFavouriteList(long userId) throws UserNotFoundException;
+	public Long[] getFavouriteList(long userId) throws UserNotFoundException;
 
 	/**This method adds a favourite to the list of a user
 	 * @param userId the id of the user
@@ -28,7 +28,7 @@ public interface UserManager extends UserDetailsService{
 	 * @throws UserNotFoundException if no user with the id userId exists
 	 * @throws FavouriteAlreadyExistsException if a favourite with the id listingId is already in the database
 	 */
-	public void addFavourite(long userId, int listingId) throws UserNotFoundException, FavouriteAlreadyExistsException;
+	public void addFavourite(long userId, long listingId) throws UserNotFoundException, FavouriteAlreadyExistsException;
 	
 	/**This method removes a favourite from the list of a user
 	 * @param userId the id of the user
@@ -58,10 +58,10 @@ public interface UserManager extends UserDetailsService{
 	public User[] getAllUsers();
 
 	/**This method sets the isAdmin-boolean of a user
-	 * @param userId the id of the user
+	 * @param l the id of the user
 	 * @param true, if the user should be an admin
 	 * @throws UserNotFoundException if no user with the id userId exists
 	 * @throws UserHadAlreadyTheRightAdminStatusException if the boolean was already in the same state as the parameter shouldBeActive
 	 */
-	public void setIsAdminOnUser(int userId, boolean shouldBeAdmin) throws UserNotFoundException, UserHadAlreadyTheRightAdminStatusException;
+	public void setIsAdminOnUser(long userId, boolean shouldBeAdmin) throws UserNotFoundException, UserHadAlreadyTheRightAdminStatusException;
 }
